chore: clean up cypress folder (#2687)
None of the server code is in use.
