+++ cypress/integration/edit/edit_dashboard/show_description.js
@@ -1,6 +1,6 @@
 import { When, Then } from 'cypress-cucumber-preprocessor/steps'
 import { clickViewActionButton } from '../../../elements/viewDashboard.js'
-import { getApiBaseUrl } from '../../../support/server/utils.js'
+import { getApiBaseUrl } from '../../../support/utils.js'
 
 const SHOW_DESC_RESP_CODE_SUCCESS = 201
 const SHOW_DESC_RESP_CODE_FAIL = 409
