+++ cypress/integration/view/view_dashboard/toggle_show_more_dashboards.js
@@ -4,8 +4,7 @@ import {
     dashboardsBarContainerSel,
     showMoreLessSel,
 } from '../../../elements/viewDashboard.js'
-import { getApiBaseUrl } from '../../../support/server/utils.js'
-import { EXTENDED_TIMEOUT } from '../../../support/utils.js'
+import { getApiBaseUrl, EXTENDED_TIMEOUT } from '../../../support/utils.js'
 
 const MIN_DASHBOARDS_BAR_HEIGHT = 71
 const MAX_DASHBOARDS_BAR_HEIGHT = 431
