--- cypress/support/server/constants.js
@@ -1,9 +1,0 @@
-export const API_STUB_MODES = {
-    STUB: 'STUB',
-    DISABLED: 'DISABLED',
-    CAPTURE: 'CAPTURE',
-}
-export const DEFAULT_API_STUB_MODE = API_STUB_MODES.STUB
-
-export const NETWORK_FIXTURES_FILE_PATH =
-    'cypress/fixtures/network/requests.json'
