--- cypress/support/server/index.js
@@ -1,1 +1,0 @@
-export { default as enableNetworkShim } from './enableNetworkShim.js'
